Move downloading code out of executable invoker to a shareable place
